requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
biopython==1.71
certifi==2018.8.24
chardet==3.0.4
click==6.7
confindr==0.4.7
COWBAT==0.4.4.5
cycler==0.10.0
daiquiri==1.3.0
decorator==4.2.1
diversitree==0.0.1
geneseekr==0.3.1
git-pull-request==2.4.0
idna==2.7
interop==1.1.4
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.11.1
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
matplotlib==2.2.2
more-itertools==4.3.0
numpy==1.14.2
olcnastools==1.1.5
OLCTools==0.7.0
ordered-set==2.0.2
pandas==0.22.0
parso==0.1.1
pathlib2==2.3.2
pexpect==4.4.0
pickleshare==0.7.4
pluggy==0.7.1
prompt-toolkit==1.0.15
psutil==5.4.7
psycopg2==2.7.4
ptyprocess==0.5.2
py==1.6.0
PyGithub==1.35
Pygments==2.2.0
PyJWT==1.5.3
PyLaTeX==1.3.0
pyparsing==2.2.0
pysam==0.15.1
pytest==3.8.2
python-dateutil==2.6.1
python-redmine==2.0.2
pytz==2017.3
pyzmq==17.0.0
rauth==0.7.3
requests==2.19.1
scipy==1.0.0
seaborn==0.9.0
simplegeneric==0.8.1
sipprverse==0.2.16
six==1.11.0
SQLAlchemy==1.2.4
tornado==5.0.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
xlrd==1.1.0
XlsxWriter==1.0.2
